Revise design for optional features and AOT
Revise the design for handling optional kernel features for AOT
compilations.  There is no need to create a separate tool
"aspect-filter" because the filtering can be done in the
"sycl-post-link" tool instead.  This is better aligned with our
long term direction because "sycl-post-link" will need to do some
AOT-specific IR transformations in the future as part of the design
presented in #9127, and those transformation will also make use of
the device configuration file.

This commit also addresses some issues that were missed before related
to embedding several AOT-compiled offload bundles into the host
executable.