[CI] Make build artifact archive name/compression parameters
I plan to enable making releases from Github CI tasks and we'll need to
use gzip compression instead of zstd as well as provide distinct names
for Linux/Windows.

An alternative would be to unpack/repack in the release action, but the
solution implemented in this PR seems to be clearer.