[SYCL] Optimize stream E2E test
Defining `SYCL_SIMPLE_SWIZZLES` significantly increases compilation time
of the test (16s -> 1m 16s on my machine) and we don't really need to
have all those swizzles in the test. We could use other methods, which
return `__swizzled_vec__`, such as `odd` or `hi`, which is done in this
commit.