[SYCL] Fix handling of non-packed structs in SpecConstants pass
This patch fixes a regression accidentally introduced in intel/llvm#9874
(presumably because we didn't have the right test).

SYCL-CTS produces an LLVM IR pattern, where spec constant initializer
contains less elements than there are in a spec constant type, because
some of those elements are implicitly-created paddings. The patch
updates the pass to handle this situation.

Resolves intel/llvm#10129