[XPTI] Refactoring - Removing TBB dependency (#10187)
XPTI framework provided an option to build it with std::thread or TBB,
with std::thread being the default library. However, many of the
performance and debug tests relied on TBB to check for data races and
data consistency. TBB constructs have been removed from the framework
and the tests have been reworked to use std::thread

---------

Signed-off-by: Vasanth Tovinkere <vasanth.tovinkere@intel.com>