[umf] create/destroy memory tracker in lib ctor/dtor
instead of relying on global variable init/destruction on
linux and on windows (when building UMF as a shared lib).

In C++ no guarantee is made for initialization order of
static variables across translation units.

This is problematic when an application (e.g. SYCL) wants
to do some cleanup in it's library destructor. Such application
might want to specify priority for it's destructor (to make sure
it's executed before any other library desturctor) but it doesn't
seem to work for static variables.

Memory tracker is still created as a global object when building
UMF as a static lib on Windows - there is no
__attribute__((destructor)) equivalent that would work for static
library.