[SYCL-MLIR] Raise `sycl.handler.set_nd_range` constructs
Detect arguments specifying the nd-range of a `parallel_for`
invocation.

Raising of calls receiving an `nd_range` fail to be raised as we are
not yet raising such constructors. Raising of calls receiving both the
global size and an offset fail because we cannot detect the
constructor. We will need to implement a pattern raising `nd_range`
constructors to fix the former and use `ReachingDefinitionAnalysis`
in the pattern implemented by this commit to fix the later.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>