[SYCL-MLIR]: Extend reaching def analysis to work inter-procedurally (#10325)
`ReachingDefinitionAnalysis` needs to use an auxiliary analysis
`UnderlyingValueAnalysis` (introduced by this patch) to work properly
inter-proceduraly. The `UnderlyingValueAnalysis` is used to connect the
reaching defs of a caller with the reaching defs of the callee. For
example, given:
```
func.func private @callee(%ptr: memref<i32>) -> memref<i32> {
  return {tag = "callee_ptr"} %ptr : memref<i32>
}

func.func @caller() {
  %ptr = memref.alloc() : memref<i32>
  %c0 = arith.constant 0 : i32
  memref.store %c0, %ptr[] {tag_name = "caller_store"} : memref<i32>
  %0 = func.call @callee(%ptr) : (memref<i32>) -> memref<i32>
  return
}
```
The reaching def for `%ptr` in `caller` is "caller_store", while the
reaching def for `%ptr` in `callee` is the initial def we give to each
function parameter. The "underlying value" of `%ptr` in `callee` is the
value yielded by ` memref.alloc()` in `caller`, and that allows the
reaching def analysis to determine that the reaching def of the `%ptr`
parameter is "caller_store".

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>