[SYCL] Make image constructor notification pass reference optional
On some build configurations, the std::optional<image_target> used in
sampledImageConstructorNotification and
unsampledImageConstructorNotification may not get correctly passed to
the library, causing a segfault when accessed in any way. This is fixed
by making it a constant reference instead.

Note: This changes only recently added symbols, so it is not a
release-to-release ABI break.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>