[SYCL][doc][NFC] Update doc and usage of -fsycl-device-obj
The -fsycl-device-obj option replaces the -fsycl-use-bitcode option.
The -fsycl-use-bitcode option is deprecated as well.  Update the
documentation to reflect this fact, also update any testing using the
older option with a replacement of the new option.

The use of -fsycl-use-bitcode should also emit the proper deprecation
message when used on the command line.