[CI] Allow llvmspirv_pulldown branch in pre-/post-commit triggers (#10439)
For post-commit it would start the entire task whenever the branch is
updated. For windows pre-commit it would allow the temporary workaround
of creating a fake PR into that branch to trigger Windows pre-commit
task (same as was done in #10434)

SYCL Nightly doesn't have "push:" trigger and it has a bug with pushing
the containers even for branches that aren't "origin/sycl" so not
changing it for now.