[SYCL][NFC] Clean-up usage of "static inline" functions (#10378)
For member functions inline which are defined in class scope inline is
implicit and can be dropped.
For free functions static causes compiler to duplicate the body of the
function when the function is not inlined and linked with another TU
defining the same function. Using inline in this case allows compiler to
merge definitions from different TU and avoid code bloat.