[SYCL][Graph] Removed outdated test
The `RecordReplay/concurrent_graph.cpp` test is not valid as
it is intending to check that an exception is thrown when
`begin_recording()` is called on a different queue while a
graph is already recording. However, this is allowed by the
specification, hence why the test was XFAIL.

The BeginEndRecording unittest already comprehensively
tests this behaviour, so this test can be removed rather
than modified for correctness https://github.com/reble/llvm/blob/sycl-graph-develop/sycl/unittests/Extensions/CommandGraph.cpp#L156

This was passing unexpectedly on Windows because we were throwing
an exception. I suspect because `Queue2` was getting created with
a different context than `Queue`, which we've seen before can happen
on Windows https://github.com/reble/llvm/pull/249

closes https://github.com/reble/llvm/issues/268