Convert tests to use opaque pointers.
Most of the changes in these tests are adjusting the way test checks
work to match the output in opaque pointer mode (in particular, opaque
struct names aren't getting generated anymore).

In the case of cl_types.ll and spirv_types.ll, the tests are deleted
entirely because they would require much more invasive changes to keep
checking the same thing (since the opaque struct names aren't generally
available, and all of the types would have to be used more directly in
an intrinsic name to be available for translation.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/b6207b8