[SYCL][NFC] Clean-up includes in DPC++ headers (#10443)
I tried to remove as much as possible unnecessary includes from
sycl/queue.hpp. While I was looking at the headers included into
queue.hpp, I cleaned-up their includes as well. For instance, I noticed
that we abuse stl.hpp file, which is intended to extend STL
functionality rather than to be a common file to include a bunch of STL
headers.

Another thing we should do in the future is to re-organize the "info
descriptors" header, which today includes descriptors for all SYCL
classes. For instance, there is not possible to include only device info
descriptors.