[SYCL][NFC] Refactor includes in DPC++ headers (#10458)
The main goal of this refactoring is to reduce amount of includes to the
context.hpp header.

Two new includes are added:
1. sycl/include/sycl/detail/impl_utils.hpp is intended to replace
sycl/include/sycl/detail/common.hpp when only interopability with "impl"
class is needed. common.hpp include is overloaded with a bunch of other
functionality.
2. sycl/include/sycl/async_handler.hpp separates async_handler definiton
from exception.hpp.