[ur] Match null DDI table entry error code
When validation is disabled and an adapter does not sets a DDI table
entry to null the loader returns `UR_RESULT_ERROR_UNINITIALIZED` however
when validation is enabled `UR_RESULT_ERROR_FEATURE_UNSUPPORTED` is
returned. This patch aligns the error codes return by the loader and
validation layer when a null DDI table entry is found.