[SYCL][New PM] Move code in sycl-post-link to new pass manager (#10455)
In the sycl-post-link tool, most of the changes towards moving the
passes to use the new pass manager were added as part of this PR by
Konstantin Bobrovsky: https://github.com/intel/llvm/pull/6835
There was one site where legacy pass manager was used. This change moves
that to use the new pass manager.

It is sufficient to check for regression in existing tests. There is no
need to add a new test.

Thanks
Sincerely

---------

Signed-off-by: Arvind Sudarsanam <arvind.sudarsanam@intel.com>