[SYCL-MLIR][ArgumentPromotion] Privatize kernel wrapper (#10472)
We found a blocker for the LoopInternalization pass. We need to prove
that the loop it needs to optimize is uniform. Uniformity analysis fails
to prove that an argument passed to the kernel body function is uniform
because the gpu.func kernel calls a wrapper function which does not have
private linkage. The analysis assumes there may be other unknown calls
to that function, and so it needs to assume that some unknown caller may
pass a non-uniform argument.

---------

Signed-off-by: Tsang, Whitney <whitney.tsang@intel.com>