[Driver][SYCL][FPGA] Improve AOCX archive behaviors with -fsycl-link=image
Use of -fsycl-link=image allows for FPGA users to create a full device
binary and store it in an archive to be used later.  This archive can
then be picked up and linked in.  The information stored in the archive
was not sufficient, causing the device binary to not be 'visible' when
linked in.

Improve the situation here by saving all of the information provided by
the wrapping step (including symbols) in the archive, but also be sure
that this entire final binary is saved as it would be if we were
performing a full E2E compilation.