[SYCL][L0] Remove not needed mutex
collectEventsForReleaseAndDestroyPiZeEventList is the only
place where mutex is currently used.
This function is always called with event that own the
waitlist mutex taken.
Therefore the mutex is redundant and causes overhead.
With this change we save 2 mutex calls per event cleanup in
in-order-queue scenarios.
This means that when we clean while reaching the
ImmCmdListsEventCleanupThreshold, we save 2000 mutex
taking calls, which improves host overhead.

Once the mutex taking is removed, there are no other code
places where mutex is used, so removing it from codebase.

Signed-off-by: mmrozek <michal.mrozek@intel.com>