[compiler] Add a pass to replace target extension types
The target extension types introduced in LLVM 17 pose a problem for us
because no upstream backend supports the "spirv" types and will crash
when lowering them. This means we must replace them across the module
with "real" types.

This commit introduces a new pass and extension to the BuiltinInfo to
provide the target with a way to customize how target extension types
are removed from the module by replacing them with other types.

Events and samplers are replaced with `i32` types, and images are
replaced with a pointer to `MuxImage` (which is just a mux-friendly
rename of the undocumented `struct.Image` structure the
ImageArgumentSubstitutionPass currently uses).

Note that functions are replaced in-place (i.e., without wrappers) as a
wrapper would still keep the types lingering around in the module.

The replacement of each of three different types can be opted out, as in
reality we will want to replace images & samplers at a different time to
events.

This pass only does something in LLVM 17+, since before that point there
will never be any target extension types to replace. The pass is also
not yet scheduled in the pipeline as it's not officially ready to use -
subsequent passes still need work to provide a more stable experience.