[SYCL-MLIR] Raise `sycl.host.set_captured`. (#10449)
Adds a pattern that detects stores (or memcpy ops) to specific elements
in a kernel lambda object, and raises them to individual
`sycl.host.set_captured` ops. The lambda object is found through an
`llvm.intr.var.annotation` with the tag `kernel`. The approach for
finding assignments-of-interest is detecting GEPs of the form `<lambda
obj>[0, <capture idx>]`.

---------

Signed-off-by: Julian Oppermann <julian.oppermann@codeplay.com>
Co-authored-by: vic <victor.perez@intel.com>
Co-authored-by: Whitney Tsang <whitney.uwaterloo@gmail.com>