[SYCL][UR][L0] Fix use of large allocations
By default, L0 adapter now allows for the use of larger-than-4GB
allocations. This is done by using always the relaxed extension
in L0 and passing the proper ZE compilation option to disable
stateful optimizations. Both are needed on pre-PVC GPUs for
correctness, but may hinder performance.

This can be disabled with SYCL_PI_LEVEL_ZERO_USE_LARGE_ALLOCATIONS=0,
which restores the stateful optimizations, but does not allow
the use of larger-than-4GB allocations.

Signed-off-by: Jaime Arteaga <jaime.a.arteaga.molina@intel.com>