[SYCL][L0] Optimize mutex acquisition.
CleanupEventListFromResetCmdList sets the event as completed under mutex.
Then releases the mutex and calls CleanupCompletedEvent.
In this method the same mutex is taken again.
This causes that we take the same mutex twice for every event in
CleanupEventListFromResetCmdList.
This change adds paramter to CleanupCompletedEvent and moves the code
from CleanupEventListFromResetCmdList to CleanupCompletedEvent, which
avoids taking the same mutex twice.
This results in 1k less mutex taken when we reach
ImmCmdListsEventCleanupThreshold.
This results in 1 less mutex taken per each event,
when we clean all events on wait calls.

Signed-off-by: mmrozek <michal.mrozek@intel.com>