[CI] Make nightly builds with Github Actions CI (#10560)
They way release action works is that it overrides the release if ran
multiple times but doesn't update the tag. As such, manual workflow runs
could present a challenge. To deal with that, I also use short commit
hash as a suffix to ensure tag and release are in sync unless the
triggering event is schedule (which only runs once a day and doesn't
result in overwrite issue).