[SYCL] Fixup SYCL metadata when dead argument is optimized and enable DAE for ESIMD
This change enables DAE for ESIMD kernels and adds code to fixup metadata inserted
by the SYCL FE when an argument is removed.

The metadata fixup is required because downstream code in llvm-spirv assumes
the number of operands for the metadata will match the actual number of arguments
for the kernel.

Another option would be to fix the downstream code, but making the metadata be accurate
to the actual kernel the downstream code will see at the point when the kernel signature changes
seemed to be a better root cause fix to me.

Let me know if you disagree with the above, or if you have a way to prevent explicitly
listing all of the metadata to fix.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>