[SYCL] vec abi unification and trivially copyable (#9492)
adapted from https://github.com/intel/llvm/pull/6816   

sycl::vec now uses std::array on host in all cases, but device access is
via `ext_vector_type` so as to not lose performance. Alignment is set
the same as the size to a maximum of 64 bytes, which aligns with an
upcoming spec change proposal. In the cases where the size is greater
than 64 we use std::array on the device. We no longer need to emit a
pragma on Windows and we can use `alignas()` on all platforms.