Fix delete of functions that becomes unused (#2109)
After the first loop of deleting instructions in ValuesToDelete, deleted
instructions in ValuesToDelete are in an unstable state. Then in the
second loop of deleting, dyn_cast to GlobalValue could return true for
an instruction and double eraseFromParent causes crash.

Global values in ValuesToDelete are functions. Unused functions are
deleted by eraseUselessFunctions anyway.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/aea1ac7