[SYCL-MLIR][sycl-constant-propagation] Handle accessor arguments (#10627)
`sycl::accessor` kernel arguments are flattened before being passed to a
kernel, enabling some constant propagation possibilities here. Each
accessor is flattened to four arguments: pointer, access range, memory
range and offset. The following rules are implemented to handle these
arguments:

- Offset and access range arguments can be replaced by a constant if
these are constants in the code;
- Offset can be replaced with the default offset (all zeroes) if no
offset is passed when building the accessor parameter;
- Memory range can be used in place of the access range if no range is
passed when building the accessor. If the memory range is known, both
can be replaced with such constant.

In addition to this, two changes need to be done to the pass:

- `ConstantExplicitArg` now also needs to track "hits", as
`ConstantAccessorArg` could replace more than a single argument;
- As an argument in `sycl.host.schedule_kernel` can represent more than
one kernel arguments, a `getTrueIndex` function is introduced to obtain
the true index (kernel argument index) of a `sycl.host.schedule_kernel`
parameter.

---------

Signed-off-by: Victor Perez <victor.perez@codeplay.com>