[SYCL][InvokeSimd] Allow callables to return uniform
The spec states that returning a `uniform` object is allowed:

"Return values of type sycl::ext::oneapi::experimental::uniform<T> are not anyhow converted, and broadcast to each work-item;
every work-item in the sub-group receives the same value.
NOTE: sycl::ext::oneapi::experimental::uniform<T> return type is the way to return a uniform value of simd or simd_mask type."

Update the compile-time error checking and ESIMD verifier to allow this.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>