[SYCL][ESIMD] Remove LowerESIMDVecArg pass
After https://github.com/intel/llvm/commit/0c51f6082760f15d268bb4320c832cebe91704e9, opaque pointers are enabled by default for device code.

This means that LowerESIMDVecArg is now never run because we do a runtime check in sycl-post-link and only run it if opaque pointers are off.

Remove the pass since it's now dead code, and remove related tests. The tests check for changes unique to LowerESIMDVecArg and cannot be ported
to instead use the pass that is run when opaque pointers is enabled, ESIMDOptimizeVecArgCallConvPass, and that pass has its own tests already.

For the tests removed under sycl/test, all of those tests already have opaque pointer versions in-tree with the _opaque suffix, so there is no
test coverage loss.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>