[SYCL][Graph] Implement graph creation and finalization for CUDA backend (#279)
* [SYCL][Graph] Implement graph creation and finalization for CUDA backend

Implements CommandBuffer creation, retain, release and finalize functions
for the cuda backend.
These functions rely on the cuda graph feature which is part of cuda runtime.
Consequently, this PR adds the required linking dependencies.
Fixes a commandbuffer bug in the cuda pluging initialization.
Reports cuda backend as supported for the graph extension.
Adds a tests that creates and finalizes an empty graph.