Don't wrap kernels that are not being called in the module (#2119)
* Don't wrap kernels that are not being called in the module

This patch is a result of a reflection about previously merged PR
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/pull/1149
"add an entry point wrapper around functions (llvm pass)"
and is enspired by various reported translator, clang (OpenCL) and Intel
GPU drivers issues (see
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/issues/2029 for
reference).

While SPIR-V spec states:
===
*OpName*
--//--. This has nosemantic impact and can safely be removed from a module.
===
yet having EntryPoint function and a function that shares the name via
OpName might be confusing by both (old) drivers and programmers, who
read the SPIR-V file.

This patch prevents generation of the wrapper function when it's not
necessary to generate it aka if a kernel function is not called by other kernel.

We can do better in other cases as well, for example I have experiments
of renaming a wrapped function adding a previx, so it could be
distinguished from the actual kernel/entry point, but for now it doesn't
pass validation for E2E OpenCL tests.

Signed-off-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>

* prevent a copy

Signed-off-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>

This patch is a result of a reflection about previously merged PR #1149 "add an entry point wrapper around functions (llvm pass)" and is enspired by various reported translator, clang (OpenCL) and Intel GPU drivers issues (see

While SPIR-V spec states:
OpName
--//--. This has nosemantic impact and can safely be removed from a module.
yet having EntryPoint function and a function that shares the name via OpName might be confusing by both not-up-to-date drivers and programmers, who read the SPIR-V file.

This patch prevents generation of the wrapper function when it's not necessary to generate it aka if a kernel function is not called by other kernel.

Signed-off-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/46285e4