[SYCL] Skip alloca commands when checking for leaves completion (#10740)
Scheduler::checkLeavesCompletion checks status of all leaves of the
buffer to see whether we can destroy that sycl::buffer. There are many
scenarios when alloca commands are leaves, these commands don't have
associated event and currently they are always incorrectly considered
"in progress" because of that preventing buffers to be destroyed timely
and deferring their destruction till the point of program termination.
Skip alloca commands to fix that.