[SYCL] Disable PI unit tests by default
These tests are currently disabled in the CI and ended up broken.

Since we're also currently moving to the Unified Runtime we might be
able to replace most of them with the Unified Runtime CTS, or they will
also need to be ported to the Unified Runtime.

So disable them by default until we figure out a proper course of action
of these.