[SYCL] Fix return type of nan builtin
This commit fixes the return type of the nan builtin. After the recent
refactoring it had been changed to always return a floating point type
with the same byte-size as the operand, but on Windows this means that
long returns a float which does not correspond to the return type
mandated by the spec.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>