[SYCL][Graph] Duplicate sub-graph nodes (#292)
* [SYCL][Graph] Duplicate sub-graph nodes

Duplicates the nodes of a sub-graph when added to its parents to enable multiple parents.
As a result nodes of the initial sub-graph and the one of the main graph are no longer
the exact same (shared_pointer) but two different nodes with similar content.
A few unitests have been updated accordingly and an equal operator has been added
to node_impl to ease comparison between nodes.