[SYCL][COMPAT] Removing Gtests from syclcompat tests
This commit reverts the changes introduced in #9976 to sycl/unittests.
It also removes the gtest suite from syclcompat

Rationale: We need device code to test the following syclcompat PRs.

Changes:
- Removed sycl/cmake/modules/AddSYCLLibraryUnittest.cmake as it was only
used here, this has also introduced overhead to people working in the
Unittests CI, it is better for them to remove the need to maintain this.
- Reverted syclcompat changes in sycl/unittests/CMakeLists.txt
- Refactored sycl/unittests/syclcompat to sycl/test-e2e/syclcompat