[SYCL][CUDA] Change builtin selection for SYCL (#9768)
Libdevice for NVPTX was previously working due to the fact that LLVM
intrinsics were not selected due to the CUDA toolchain having
isMathErrnoDefault evaluate to `true`. Since LLVM intrinsics were not
selected then the symbols could be found when linking with libdevice,
which gives special backend specific definitions of CXX stdlib funcs.

Using errno to prevent intrinsic selection was a hack and it gets undone
by using `-ffast-math`, meaning libdevice CXX funcs were not working
with `-ffast-math`.

This change instead explicitly says not to use LLVM intrinsics if
compiling SYCL for NVPTX backend. This means that `-ffast-math`
behaviour should now be fixed for CXX stdlib funcs.

@jchlanda