[SYCL] Use boost's unordered maps for kernel caching (#10480)
Fetching the kernel from the map is a hotspot for applications that
repeatedly launch microkernels. As the number of kernels used by the
application grows, std::map shows poor performance there.

This patch replaces uses of std::map with boost::unordered_flat_map and
boost::unordered_map. The former is used for the fast path to prioritize
lookup performance, the latter is used for the slow path since it's
expected to be used only once for each newly created kernel, so the
higher insertion overhead of the flat map is more of an issue there.