[SYCL] annotated_ptr API fixes (#10124)
1) annotated_ptr constructor is made explicit (removed - will be
seperate PR)
2) annotated_ref conversion constructor is removed
3) annotated_ref assignment returns by value
4) annotated_ref assignment never rebinds but assigns
5) annotated_ref copy constructor is deleted
6) annotated_ref isn't supposed to in an anonymous namespace
7) annotated_ref add incr/decr operators and clarify compound operators

Justification:
1) The constructors were not intended as conversion constructors. They
can take a single argument (for the first constructor the 2nd argument
is defaulted and for the second the pack can be empty), and therefore
can act as conversion constructor.
2) This was noticed before
https://github.com/intel/llvm/pull/8270#discussion_r1102035230 but not
resolved
3) Returning a reference would mean that for `a=b=2`, `b` needs to be
loaded after it is stored, if a special load/store functions are used.
We don't want that. Instead we follow the precedent set by `atomic_ref`
and return by value.
4) Since the initial version `operator=(T)` did value assignment.
`operator=(annotated_ref<T,...>)` needs to do the same and shouldn't
rebind. Otherwise `atomic_ptr<int, ...> a...; a[0] = a[1];` would rebind
the temporary ref and not assign.
5) `operator=` and copy constructor should behave consistent. And
because `operator=` assigns the copy constructor shouldn't bind.

Warning:
Number 4 is a change of behavior. Code using it will change behavior and
not see a compiler diagnostic.

---------

Co-authored-by: Chen, Brox <brox.chen@intel.com>