[cmake] Ensure code is compiled with -Werror in dev mode
The recent hotfix in #800 requiring the v0.7.1 tag was fixing an
instance of code not compiling when `-Werror` is specified. We do
already specify `-Werror` with `UR_DEVELOPER_MODE=ON` in the GitHub
Actions, however the `ur_common` library was not being compiled
explicitly with this flag.

As setup prior to this patch `ur_common` was an interface library. This
means the source files are not compiled into their own static library
but instead included in any targets which "link" it.

This patch changes the `ur_common` into a static library that is always
compiled into a static library avoiding recompiling the source files in
each target which makes use of it. Adapters will now link with the
static library instead of including its source files into their own
targets build steps. Additionally, the test suite `test-loader-config`
was also updated to use `add_ur_exectuable` and thus respect the flags
set when `UR_DEVELOPER_MODE=ON`.