[SYCL] Use dummy output variables to check piGetDeviceAndHostTimer support
MDeviceHostBaseTime data member of device_impl is supposed to be properly
updated/calculated using getCurrentDeviceTime(). Currently it is also
used to check piGetDeviceAndHostTimer support in
isGetDeviceAndHostTimerSupported() which results in wrong values stored
in MDeviceHostBaseTime and wrong calculation of submit time.
That's why Use temp variables to check support of
piGetDeviceAndHostTimer to avoid spoiling MDeviceHostBaseTime.