[SYCL] Don't spoil MDeviceHostBaseTime during isGetDeviceAndHostTimerSupported (#10909)
`MDeviceHostBaseTime` data member of `device_impl` is supposed to be
properly updated/calculated using `getCurrentDeviceTime()`. Currently it
is also used to check `piGetDeviceAndHostTimer` support in
`isGetDeviceAndHostTimerSupported()` which results in wrong values
stored in `MDeviceHostBaseTime` and wrong calculation of submit time.
That's why use temp variables to check support of
`piGetDeviceAndHostTimer` to avoid spoiling `MDeviceHostBaseTime`.