[SYCL][COMPAT] Added device extension and manager (#10774)
This is the second PR implementation of SYCLcompat. SYCLcompat primary
goals and features are documented
[here](https://github.com/intel/llvm/tree/sycl/sycl/doc/syclcompat).

The PR includes the code that would implement the wrapper to expose the
management of the `sycl::device`, `sycl::queue`, and `sycl::context`.
The exposed functionalities include:
- Creation and destruction of queues, and providing the ability to wait
for submitted kernels.
- Any async errors will be output to stderr if
`print_on_async_exceptions`. Synchronous exceptions have to be managed
by users independently of what is set in this parameter.
- Management of multiple devices. The library uses the default SYCL
device (`sycl::default_selector_v`) as the default device, and exposes
all other devices available on the system through the
`syclcompat::select_device`member function.

We are open to any suggestions, concerns, or improvements you may have,
so please, let us know if you have any.

---------

Co-authored-by: Gordon Brown <gordon@codeplay.com>
Co-authored-by: Joe Todd <joe.todd@codeplay.com>
Co-authored-by: Pietro Ghiglio <pietro.ghiglio@codeplay.com>
Co-authored-by: Ruyman Reyes <ruyman@codeplay.com>