[SYCL][Driver] Fix fno-gpu-rdc SYCL libraries linking not using opaque pointers (#10418)
Fix fno-gpu-rdc SYCL libraries linking not using opaque pointers for
non-SPIR targets.

Issue was when specifiying -fno-gpu-rdc for an opaque pointer target,
there would be object modules with typed pointers. so llvm-link would
need the opaque-pointer flag but that wasn't being added previously due
to relying on the default RDC behaviour where the first call to
llvm-link linking the SYCL libraries with the kernel module is expected
to have the kernel module as the first input. This isn't true in no-rdc
mode, where the first input is a SYCL device library, hence checking the
RDC mode before setting the state of `LinkSYCLDeviceLibs`.

Test checking the behaviour was added as well.