[SYCL-MLIR] Avoid empty LLVM global op for COMDAT (#10927)
In the current version on our branch, `mlir-translate` can create empty
LLVM global operations in the MLIR output, if the input LLVM IR module
does not contain any COMDAT information (which can be omitted according
to the LLVM language reference manual). The empty global operation
causes the verifier to fail.

This problem has been fixed in upstream LLVM with [this
revision](https://reviews.llvm.org/D153739), but this fix has not yet
landed in our branch. The changes in this PR can be reverted as soon as
we receive [the revision](https://reviews.llvm.org/D153739) through
pull-down, but allow us to move forward in the meantime.