[SYCL][KernelFusion] Ensure queue_impl is not unnecessary referenced
The KernelFusionCmd keeps a references to the queue on which it was submitted.
However these commands are also kept on the side in order
to ensure event validity and are only reset when start fusion is called
again on the same queue.
This forces queue_impl and events to be kept alive as the command keeps
the last reference until the sycl shutdown procedure.

The patch forces 1) the command to reset it pointers to the queue
once enqueued and 2) to clean up the KernelFusionCmd
attached to the queue to free up events.

Signed-off-by: Victor Lomuller <victor@codeplay.com>