[SROA] Add `-opaque-pointers=1` to test that uses mixed ptr types.
Necessary after 39a41c8905618f due to community committing a test that
uses both typed pointers and opaque pointers. Since the typed pointers
appear first in the test, the parser disables opaque pointers. This
causes an error when it tries to parse the opaque pointers in the rest
of the test.

Fixes the following LIT regression:

    LLVM :: Transforms/SROA/scalable-vector-struct.ll