[SYCL]Replaced LoadLibraryA() with LoadLibraryEx() (#10914)
Replaced LoadLibraryA() with LoadLibraryEx() and added
SetDllDirectoryA() to remove the working directory from the standard
search path. LoadLibraryExA() was used because "path" is of type
c-string.

- LoadLibraryExA() <-- for ANSI (ASCII)
- LoadLibraryExW() <-- for Unicode

The libloaderapi.h header defines LoadLibraryEx as an alias which
automatically selects the ANSI or Unicode version of this function based
on the definition of the UNICODE preprocessor constant.

https://learn.microsoft.com/en-us/windows/win32/api/libloaderapi/nf-libloaderapi-loadlibraryexw

HMODULE LoadLibraryEx(
[in] LPCSTR lpLibFileName, //==> library path as a string
HANDLE hFile, //==> reserved as NULL for future use
[in] DWORD dwFlags //==> dwFlags determine the behavior of
LoadLibaryEx(). When it is NULL, it behaves same as LoadLibaryA()
);

https://learn.microsoft.com/en-us/windows/win32/api/libloaderapi/nf-libloaderapi-loadlibraryexa

---------

Signed-off-by: Kim, Euna <euna.kim@intel.com>