[SYCL][Graph] Fix issues with sycl_ext_oneapi_graph subgraphs (#10822)
This PR addresses some issues with the
[sycl_ext_oneapi_graph](https://github.com/intel/llvm/blob/sycl/sycl/doc/extensions/proposed/sycl_ext_oneapi_graph.asciidoc)
extension regarding sub-graphs. A sub-graph is when an executable
`command_graph` is added as a node in another graph.

Currently, adding an executable graph object as a subgraph permanently
connects the child graph's root nodes to the parent graph. This affects
subsequent independent submissions of the executable child graph object
and later additions as a subgraph node. This is incorrect behaviour, as
the previous uses as a sub-graph should have no effect on the state of
the executable graph object.

Fixed in this PR by duplicating the nodes of a subgraph when added to a
parent, enabling multiple parents. As a result, nodes of the initial
subgraph and the one of the main graph are no longer the exact same
(shared_pointer) but two different nodes with similar content.

## Authors

Co-authored-by: Maxime France-Pillois
<maxime.francepillois@codeplay.com>

---------

Co-authored-by: Maxime France-Pillois <maxime.francepillois@codeplay.com>
Co-authored-by: Ben Tracy <ben.tracy@codeplay.com>