[SYCL-MLIR] Fix image type handling (#11000)
With the switch to opaque pointers, image types are now represented by
LLVM TargetExtensionTypes. These are not allowed as element type of a
`memref`.

This therefore changes the handling to use `llvm.ptr` instead, to avoid
this error.

It re-enables three SYCL end-to-end tests that were put on the list of
expected failures as part of #10941.