Add SPV_EXT_image_raw10_raw12 writer support (#2132)
Add basic LLVM-to-SPIR-V support for the SPV_EXT_image_raw10_raw12
extension, enabling the extension if constants from the extension are
found in the LLVM IR.

The extension adds 2 new return values for `OpImageQueryFormat`, which
are integer constants that may appear anywhere in LLVM IR.
Distinguishing between the extension's constants and arbitrary
constants that happen to have the same value as the extension's
constants is not possible in general.  Hence this patch only covers
some common use cases, where the result of `OpImageQueryFormat` is
directly used in an integer comparison or switch instruction.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/8564fd4